pyproject.toml
[project]
name = "bim-guard-simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastlite>=0.2.4",
    "ifcopenshell>=0.8.4.post1",
    "litellm>=1.83.4",
    "monsterui>=1.0.45",
    "pandas>=3.0.2",
    "pypdf>=4.3.1",
    "python-dotenv>=1.0.1",
    "python-fasthtml>=0.12.48",
]
